requirements.txt
amqp==5.0.5
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
backcall==0.2.0
beautifulsoup4==4.9.1
billiard==3.6.3.0
blog==1.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
Django==3.0.5
django-cors-headers==3.5.0
django-crontab==0.7.1
django-phonenumber-field==5.0.0
django-rest-framework==0.1.0
djangorestframework==3.11.1
djangotoolbox==1.8.0
djongo==1.3.3
dnspython==2.0.0
docopt==0.6.2
idna==2.10
imageio==2.9.0
imutils==0.5.3
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.5.2
jedi==0.17.2
Jinja2==3.0.0a1
kiwisolver==1.3.1
kombu==5.0.2
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
mongo==0.2.0
mongoengine==0.20.0
mysqlclient==2.0.1
npm==0.1.1
optional-django==0.1.0
parso==0.7.1
phonenumbers==8.12.9
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.8
pycodestyle==2.6.0
pyembed==1.3.3
pyembed-markdown==1.1.0
Pygments==2.6.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pymdown-extensions==8.0
pymongo==3.11.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
pywin32==300
PyYAML==5.3.1
pyzmq==20.0.0
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.2.4
toml==0.10.1
tornado==6.1
traitlets==5.0.5
urllib3==1.25.10
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1
